requirements.txt
numpy
pandas
langchain
openai
langchain-openai
chromadb
python-dotenv==1.0.1
unstructured==0.12.5
unstructured-client==0.21.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12